WIP Extract search context
