+++ client/web/src/__mocks__/legacyLayoutRouteContext.ts
@@ -35,13 +35,11 @@ export const windowContextConfig = {
 export const injectedAppConfig = {} as unknown as StaticInjectedAppConfig
 
 export const staticWebAppConfig = {
-    setSelectedSearchContextSpec: () => {},
     platformContext: NOOP_PLATFORM_CONTEXT as PlatformContext,
     extensionsController: null,
 } satisfies StaticSourcegraphWebAppContext
 
 export const dynamicWebAppConfig = {
-    selectedSearchContextSpec: '',
     authenticatedUser: null,
     settingsCascade: {
         final: null,
